+++ src/java.desktop/share/native/libharfbuzz/OT/Layout/GPOS/ValueFormat.hh
@@ -64,7 +64,7 @@ struct ValueFormat : HBUINT16
   // be added to ValueFormat. As such, we use the faster hb_popcount8
   // that only processes the lowest 8 bits.
   unsigned int get_len () const  { return hb_popcount8 ((uint8_t) *this); }
-  unsigned int get_size () const { return get_len () * Value::static_size; }
+  size_t get_size () const { return get_len () * Value::static_size; }
 
   hb_vector_t<unsigned> get_device_table_indices () const {
     unsigned i = 0;
@@ -304,9 +304,9 @@ struct ValueFormat : HBUINT16
     return *static_cast<const Offset16To<Device, ValueBase> *> (value);
   }
   static inline const Device& get_device (const Value* value,
-                                          bool *worked,
-                                          const ValueBase *base,
-                                          hb_sanitize_context_t &c)
+					  bool *worked,
+					  const ValueBase *base,
+					  hb_sanitize_context_t &c)
   {
     if (worked) *worked |= bool (*value);
     auto &offset = *static_cast<const Offset16To<Device> *> (value);
